Remove dependency between Ignition and IgnitionBase
Building Ignition from IgnitionBase had little benefit in terms
of code reuse. It was also harder to maintain the unit tests
because a change in IgnitionBase may unknowning affect
Ignition unit tests. Also building an IgnitionBase dependency
is not trivial and prone to errors.

Ignition no longer builds from IgnitionBase. They do share
common functions in building their config and in their unit
tests.

IgnitionBase has been renamed to UnconfiguredIgnition.
