+++ pkg/asset/agent/image/ignition.go
@@ -1,6 +1,7 @@
 package image
 
 import (
+	"context"
 	"fmt"
 	"net"
 	"net/url"
@@ -10,6 +11,8 @@ import (
 
 	"github.com/coreos/ignition/v2/config/util"
 	igntypes "github.com/coreos/ignition/v2/config/v3_2/types"
+	"github.com/coreos/stream-metadata-go/arch"
+	"github.com/google/uuid"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
@@ -25,7 +28,10 @@ import (
 	"github.com/openshift/installer/pkg/asset/ignition/bootstrap"
 	"github.com/openshift/installer/pkg/asset/password"
 	"github.com/openshift/installer/pkg/asset/tls"
+	"github.com/openshift/installer/pkg/rhcos"
+	"github.com/openshift/installer/pkg/types"
 	"github.com/openshift/installer/pkg/types/agent"
+	"github.com/openshift/installer/pkg/version"
 )
 
 const rendezvousHostEnvPath = "/etc/assisted/rendezvous-host.env"
@@ -58,7 +64,7 @@ type agentTemplateData struct {
 	PublicContainerRegistries string
 	InfraEnvID                string
 	ClusterName               string
-	OSImage                   models.OsImage
+	OSImage                   *models.OsImage
 	Proxy                     *v1beta1.Proxy
 }
 
@@ -80,7 +86,6 @@ func (a *Ignition) Dependencies() []asset.Asset {
 		&agentconfig.AgentConfig{},
 		&mirror.RegistriesConf{},
 		&mirror.CaBundle{},
-		&IgnitionBase{},
 	}
 }
 
@@ -116,19 +121,30 @@ func (a *Ignition) Generate(dependencies asset.Parents) error {
 	agentManifests := &manifests.AgentManifests{}
 	agentConfigAsset := &agentconfig.AgentConfig{}
 	extraManifests := &manifests.ExtraManifests{}
-	ignitionBaseAsset := &IgnitionBase{}
-	dependencies.Get(agentManifests, agentConfigAsset, extraManifests, ignitionBaseAsset)
+	dependencies.Get(agentManifests, agentConfigAsset, extraManifests)
 
 	pwd := &password.KubeadminPassword{}
 	dependencies.Get(pwd)
 	pwdHash := string(pwd.PasswordHash)
 
 	infraEnv := agentManifests.InfraEnv
-	config, err := ignitionBaseAsset.CopyIgnitionConfig()
-	if err != nil {
-		return err
+
+	config := igntypes.Config{
+		Ignition: igntypes.Ignition{
+			Version: igntypes.MaxVersion.String(),
+		},
+		Passwd: igntypes.Passwd{
+			Users: []igntypes.PasswdUser{
+				{
+					Name: "core",
+					SSHAuthorizedKeys: []igntypes.SSHAuthorizedKey{
+						igntypes.SSHAuthorizedKey(infraEnv.Spec.SSHAuthorizedKey),
+					},
+					PasswordHash: &pwdHash,
+				},
+			},
+		},
 	}
-	config.Passwd.Users[0].PasswordHash = &pwdHash
 
 	nodeZeroIP, err := RetrieveRendezvousIP(agentConfigAsset.Config, agentManifests.NMStateConfigs)
 	if err != nil {
@@ -138,13 +154,33 @@ func (a *Ignition) Generate(dependencies asset.Parents) error {
 	logrus.Infof("The rendezvous host IP (node0 IP) is %s", nodeZeroIP)
 
 	a.RendezvousIP = nodeZeroIP
+	// Default to x86_64
+	archName := arch.RpmArch(types.ArchitectureAMD64)
+	if infraEnv.Spec.CpuArchitecture != "" {
+		archName = infraEnv.Spec.CpuArchitecture
+	}
 
-	releaseImageList, err := releaseImageList(agentManifests.ClusterImageSet.Spec.ReleaseImage, ignitionBaseAsset.archName)
+	releaseImageList, err := releaseImageList(agentManifests.ClusterImageSet.Spec.ReleaseImage, archName)
 	if err != nil {
 		return err
 	}
 
-	a.CPUArch = *ignitionBaseAsset.osImage.CPUArchitecture
+	registriesConfig := &mirror.RegistriesConf{}
+	registryCABundle := &mirror.CaBundle{}
+	dependencies.Get(registriesConfig, registryCABundle)
+
+	publicContainerRegistries := getPublicContainerRegistries(registriesConfig)
+
+	releaseImageMirror := mirror.GetMirrorFromRelease(agentManifests.ClusterImageSet.Spec.ReleaseImage, registriesConfig)
+
+	infraEnvID := uuid.New().String()
+	logrus.Debug("Generated random infra-env id ", infraEnvID)
+
+	osImage, err := getOSImagesInfo(archName)
+	if err != nil {
+		return err
+	}
+	a.CPUArch = *osImage.CPUArchitecture
 
 	clusterName := fmt.Sprintf("%s.%s",
 		agentManifests.ClusterDeployment.Spec.ClusterName,
@@ -155,12 +191,12 @@ func (a *Ignition) Generate(dependencies asset.Parents) error {
 		agentManifests.GetPullSecretData(),
 		releaseImageList,
 		agentManifests.ClusterImageSet.Spec.ReleaseImage,
-		ignitionBaseAsset.releaseImageMirror,
-		ignitionBaseAsset.hasMirrorConfig,
-		ignitionBaseAsset.publicContainerRegistries,
+		releaseImageMirror,
+		len(registriesConfig.MirrorConfig) > 0,
+		publicContainerRegistries,
 		agentManifests.AgentClusterInstall,
-		ignitionBaseAsset.infraEnvID,
-		ignitionBaseAsset.osImage,
+		infraEnvID,
+		osImage,
 		infraEnv.Spec.Proxy)
 
 	err = bootstrap.AddStorageFiles(&config, "/", "agent/files", agentTemplateData)
@@ -173,28 +209,46 @@ func (a *Ignition) Generate(dependencies asset.Parents) error {
 		getRendezvousHostEnv(agentTemplateData.ServiceProtocol, nodeZeroIP))
 	config.Storage.Files = append(config.Storage.Files, rendezvousHostFile)
 
+	err = addBootstrapScripts(&config, agentManifests.ClusterImageSet.Spec.ReleaseImage)
+	if err != nil {
+		return err
+	}
+
 	// add ZTP manifests to manifestPath
 	for _, file := range agentManifests.FileList {
 		manifestFile := ignition.FileFromBytes(filepath.Join(manifestPath, filepath.Base(file.Filename)),
 			"root", 0600, file.Data)
-		config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, manifestFile)
+		config.Storage.Files = append(config.Storage.Files, manifestFile)
 	}
 
 	// add AgentConfig if provided
 	if agentConfigAsset.Config != nil {
 		agentConfigFile := ignition.FileFromBytes(filepath.Join(manifestPath, filepath.Base(agentConfigAsset.File.Filename)),
 			"root", 0600, agentConfigAsset.File.Data)
-		config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, agentConfigFile)
+		config.Storage.Files = append(config.Storage.Files, agentConfigFile)
 	}
 
 	addMacAddressToHostnameMappings(&config, agentConfigAsset)
 
-	err = bootstrap.AddSystemdUnits(&config, "agent/systemd/units", agentTemplateData, getEnabledServices(config))
+	err = addStaticNetworkConfig(&config, agentManifests.StaticNetworkConfigs)
+	if err != nil {
+		return err
+	}
+
+	enabledServices := getDefaultEnabledServices()
+	// Enable pre-network-manager-config.service only when there are network configs defined
+	if len(agentManifests.StaticNetworkConfigs) != 0 {
+		enabledServices = append(enabledServices, "pre-network-manager-config.service")
+	}
+
+	err = bootstrap.AddSystemdUnits(&config, "agent/systemd/units", agentTemplateData, enabledServices)
 	if err != nil {
 		return err
 	}
 
 	addTLSData(&config, dependencies)
+
+	addMirrorData(&config, registriesConfig, registryCABundle)
 
 	err = addHostConfig(&config, agentConfigAsset)
 	if err != nil {
@@ -210,25 +264,52 @@ func (a *Ignition) Generate(dependencies asset.Parents) error {
 	return nil
 }
 
-func getEnabledServices(config igntypes.Config) []string {
-	var enabledServices []string
-	// find services already enabled from IgnitionBase
-	// primarily pre-network-manager-config.service
-	for _, unit := range config.Systemd.Units {
-		if unit.Enabled != nil && *unit.Enabled {
-			enabledServices = append(enabledServices, unit.Name)
+func getDefaultEnabledServices() []string {
+	return []string{
+		"agent-interactive-console.service",
+		"agent.service",
+		"assisted-service-db.service",
+		"assisted-service-pod.service",
+		"assisted-service.service",
+		"create-cluster-and-infraenv.service",
+		"node-zero.service",
+		"multipathd.service",
+		"selinux.service",
+		"install-status.service",
+		"set-hostname.service",
+		"start-cluster-installation.service",
+	}
+}
+
+func addBootstrapScripts(config *igntypes.Config, releaseImage string) (err error) {
+	// Set up bootstrap service recording
+	if err := bootstrap.AddStorageFiles(config,
+		"/usr/local/bin/bootstrap-service-record.sh",
+		"bootstrap/files/usr/local/bin/bootstrap-service-record.sh",
+		nil); err != nil {
+		return err
+	}
+
+	// Use bootstrap script to get container images
+	relImgData := struct{ ReleaseImage string }{
+		ReleaseImage: releaseImage,
+	}
+	for _, script := range []string{"release-image.sh", "release-image-download.sh"} {
+		if err := bootstrap.AddStorageFiles(config,
+			"/usr/local/bin/"+script,
+			"bootstrap/files/usr/local/bin/"+script+".template",
+			relImgData); err != nil {
+			return err
 		}
 	}
-	// add back services disabled in IgnitionBase
-	enabledServices = append(enabledServices, agentEnabledServices...)
-	return enabledServices
+	return nil
 }
 
 func getTemplateData(name, pullSecret, releaseImageList, releaseImage,
 	releaseImageMirror string, haveMirrorConfig bool, publicContainerRegistries string,
 	agentClusterInstall *hiveext.AgentClusterInstall,
 	infraEnvID string,
-	osImage models.OsImage,
+	osImage *models.OsImage,
 	proxy *v1beta1.Proxy) *agentTemplateData {
 	return &agentTemplateData{
 		ServiceProtocol:           "http",
@@ -280,13 +361,13 @@ func addStaticNetworkConfig(config *igntypes.Config, staticNetworkConfig []*mode
 	for i := range filesList {
 		nmFilePath := path.Join(nmConnectionsPath, filesList[i].FilePath)
 		nmStateIgnFile := ignition.FileFromBytes(nmFilePath, "root", 0600, []byte(filesList[i].FileContents))
-		config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, nmStateIgnFile)
+		config.Storage.Files = append(config.Storage.Files, nmStateIgnFile)
 	}
 
 	nmStateScriptFilePath := "/usr/local/bin/pre-network-manager-config.sh"
 	// A local version of the assisted-service internal script is currently used
 	nmStateScript := ignition.FileFromBytes(nmStateScriptFilePath, "root", 0755, []byte(manifests.PreNetworkConfigScript))
-	config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, nmStateScript)
+	config.Storage.Files = append(config.Storage.Files, nmStateScript)
 
 	return nil
 }
@@ -303,13 +384,13 @@ func addTLSData(config *igntypes.Config, dependencies asset.Parents) {
 	for _, ck := range certKeys {
 		for _, d := range ck.(asset.WritableAsset).Files() {
 			f := ignition.FileFromBytes(path.Join("/opt/agent", d.Filename), "root", 0600, d.Data)
-			config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, f)
+			config.Storage.Files = append(config.Storage.Files, f)
 		}
 	}
 
 	pwd := &password.KubeadminPassword{}
 	dependencies.Get(pwd)
-	config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files,
+	config.Storage.Files = append(config.Storage.Files,
 		ignition.FileFromBytes("/opt/agent/tls/kubeadmin-password.hash", "root", 0600, pwd.PasswordHash))
 }
 
@@ -319,14 +400,14 @@ func addMirrorData(config *igntypes.Config, registriesConfig *mirror.RegistriesC
 	if registriesConfig.File != nil {
 		registriesFile := ignition.FileFromBytes(registriesConfPath,
 			"root", 0600, registriesConfig.File.Data)
-		config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, registriesFile)
+		config.Storage.Files = append(config.Storage.Files, registriesFile)
 	}
 
 	// This is required for the agent to run the podman commands to the mirror
 	if registryCABundle.File != nil && len(registryCABundle.File.Data) > 0 {
 		caFile := ignition.FileFromBytes(registryCABundlePath,
 			"root", 0600, registryCABundle.File.Data)
-		config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, caFile)
+		config.Storage.Files = append(config.Storage.Files, caFile)
 	}
 }
 
@@ -344,7 +425,7 @@ func addMacAddressToHostnameMappings(
 			file := ignition.FileFromBytes(filepath.Join(hostnamesPath,
 				strings.ToLower(filepath.Base(host.Interfaces[0].MacAddress))),
 				"root", 0600, []byte(host.Hostname))
-			config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, file)
+			config.Storage.Files = append(config.Storage.Files, file)
 		}
 	}
 }
@@ -357,7 +438,7 @@ func addHostConfig(config *igntypes.Config, agentConfig *agentconfig.AgentConfig
 
 	for path, content := range confs {
 		hostConfigFile := ignition.FileFromBytes(filepath.Join("/etc/assisted/hostconfig", path), "root", 0644, content)
-		config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, hostConfigFile)
+		config.Storage.Files = append(config.Storage.Files, hostConfigFile)
 	}
 	return nil
 }
@@ -402,11 +483,47 @@ func addExtraManifests(config *igntypes.Config, extraManifests *manifests.ExtraM
 			fileName := fmt.Sprintf("%s-%d%s", baseFileName, n, ext)
 
 			extraFile := ignition.FileFromBytes(fileName, user, mode, m)
-			config.Storage.Files = bootstrap.ReplaceOrAppend(config.Storage.Files, extraFile)
+			config.Storage.Files = append(config.Storage.Files, extraFile)
 		}
 	}
 
 	return nil
+}
+
+func getOSImagesInfo(cpuArch string) (*models.OsImage, error) {
+	st, err := rhcos.FetchCoreOSBuild(context.Background())
+	if err != nil {
+		return nil, err
+	}
+
+	osImage := &models.OsImage{
+		CPUArchitecture: &cpuArch,
+	}
+
+	openshiftVersion, err := version.Version()
+	if err != nil {
+		return nil, err
+	}
+	osImage.OpenshiftVersion = &openshiftVersion
+
+	streamArch, err := st.GetArchitecture(cpuArch)
+	if err != nil {
+		return nil, err
+	}
+
+	artifacts, ok := streamArch.Artifacts["metal"]
+	if !ok {
+		return nil, fmt.Errorf("failed to retrieve coreos metal info for architecture %s", cpuArch)
+	}
+	osImage.Version = &artifacts.Release
+
+	isoFormat, ok := artifacts.Formats["iso"]
+	if !ok {
+		return nil, fmt.Errorf("failed to retrieve coreos ISO info for architecture %s", cpuArch)
+	}
+	osImage.URL = &isoFormat.Disk.Location
+
+	return osImage, nil
 }
 
 // RetrieveRendezvousIP Returns the Rendezvous IP from either AgentConfig or NMStateConfig
